+++ function/statustwitch.php
@@ -2,71 +2,152 @@
 
 	class StatusTwitch extends Command {
 
-		public static $statusTwitch_time = 0;
-		public static $statusTwitch_online = [];
+		private static $statusTwitch_time = 0;
+		private static $statusTwitch_online = [];
+		private $token = NULL;
+		private $users = NULL;
+		private $follows = NULL;
+		private $emot = NULL;
+		private $streams = NULL;
 
+		private function setToken(): void
+		{
+			$ch = curl_init();
+				curl_setopt_array($ch, [
+				CURLOPT_URL => "https://id.twitch.tv/oauth2/token?client_id={$this->config['functions_StatusTwitch']['apikay']}&client_secret={$this->config['functions_StatusTwitch']['secret']}&grant_type=client_credentials",
+				CURLOPT_POST => 1,
+				CURLOPT_RETURNTRANSFER => true,
+			]);
+			$token = json_decode(curl_exec($ch), true);
+			if(!$token['access_token']){
+				// tu log że ni ma tokena
+			}else{
+				$this->token = $token['access_token'];
+			}
+		}
+
+		private function setUsers(string $name): void
+		{
+			$ch = curl_init();
+				curl_setopt_array($ch, [
+					CURLOPT_URL => "https://api.twitch.tv/helix/users?login={$name}",
+					CURLOPT_RETURNTRANSFER => true,
+					CURLOPT_HTTPHEADER => [
+						'Authorization: Bearer '.$this->token,
+						'Client-ID: '.$this->config['functions_StatusTwitch']['apikay']
+						],
+				]);
+			$this->users = json_decode(curl_exec($ch));
+		}
+
+		private function setFollows(): void
+		{
+			$ch = curl_init();
+				curl_setopt_array($ch, [
+					CURLOPT_URL => "https://api.twitch.tv/helix/users/follows?to_id={$this->users->data[0]->id}&first=1",
+					CURLOPT_RETURNTRANSFER => true,
+					CURLOPT_HTTPHEADER => [
+						'Authorization: Bearer '.$this->token,
+						'Client-ID: '.$this->config['functions_StatusTwitch']['apikay']
+					],
+				]);
+			$this->follows = json_decode(curl_exec($ch));
+		}
+
+		private function setEmotes(): void
+		{
+			$ch = curl_init();
+			curl_setopt_array($ch, [
+			CURLOPT_URL => "https://api.twitchemotes.com/api/v4/channels/{$this->users->data[0]->id}",
+			CURLOPT_RETURNTRANSFER => true,
+			CURLOPT_HTTPHEADER => [
+				'Authorization: Bearer '.$this->token,
+				'Client-ID: '.$this->config['functions_StatusTwitch']['apikay']
+				],
+			]);
+			$emotes = json_decode(curl_exec($ch));
+			$emot = NULL;
+			foreach($emotes->emotes as $e){
+				$emot .= "[img]https://static-cdn.jtvnw.net/emoticons/v1/{$e->id}/1.0[/img] {$e->code}\n";
+			}
+			$this->emot = $emot;
+		}
+
+		private function setStreams(string $name): void
+		{
+			$ch = curl_init();
+			curl_setopt_array($ch, [
+				CURLOPT_URL => "https://api.twitch.tv/helix/streams?user_login={$name}",
+				CURLOPT_RETURNTRANSFER => true,
+				CURLOPT_HTTPHEADER => [
+					'Authorization: Bearer '.$this->token,
+					'Client-ID: '.$this->config['functions_StatusTwitch']['apikay']
+				],
+			]);
+			$this->streams = json_decode(curl_exec($ch));
+		}
+	
 		public function execute(): void
 		{
 			if(self::$statusTwitch_time <= time()){
 				foreach($this->config['functions_StatusTwitch']['cid_name'] as $cid => $value){
-					$ch = curl_init();
-					curl_setopt_array($ch, [
-						CURLOPT_HTTPHEADER => [
-							'Accept: application/vnd.twitchtv.v5+json',
-							'Client-ID: 56o6gfj3nakgeaaqpku3cugkf7lgzk'
-						],
-						CURLOPT_RETURNTRANSFER => true,
-						CURLOPT_URL => 'https://api.twitch.tv/helix/users?login='.$value['users']
-					]);
-					$userId = json_decode(curl_exec($ch), true)['data'][0]['id'];
-					if(empty(self::$statusTwitch_online[$userId])){
-						self::$statusTwitch_online[$userId] = 0;
+					$this->setToken();
+					$this->setUsers($value['users']);
+					$this->setFollows();
+					$this->setEmotes();
+					$this->setStreams($value['users']);
+					if(empty(self::$statusTwitch_online[$this->users->data[0]->id])){
+						self::$statusTwitch_online[$this->users->data[0]->id] = 0;
 					}
-					$ch = curl_init();
-					curl_setopt_array($ch, [
-						CURLOPT_HTTPHEADER => [
-							'Accept: application/vnd.twitchtv.v5+json',
-							'Client-ID: 56o6gfj3nakgeaaqpku3cugkf7lgzk'
-						],
-						CURLOPT_RETURNTRANSFER => true,
-						CURLOPT_URL => 'https://api.twitch.tv/kraken/streams/'.$userId.'?stream_type=live'
-					]);
-					$jdc = json_decode(curl_exec($ch));
-					if(empty($jdc->stream)){
-						if(self::$statusTwitch_online[$userId] == 1){
-							self::$statusTwitch_online[$userId] = 0;
-						}
+					if(!empty($this->streams->data[0])){
 						$ch = curl_init();
-						curl_setopt_array($ch, [
-							CURLOPT_HTTPHEADER => [
-								'Accept: application/vnd.twitchtv.v5+json',
-								'Client-ID: 56o6gfj3nakgeaaqpku3cugkf7lgzk'
-							],
+							curl_setopt_array($ch, [
+							CURLOPT_URL => "https://api.twitch.tv/helix/games?id={$this->streams->data[0]->game_id}",
 							CURLOPT_RETURNTRANSFER => true,
-							CURLOPT_URL => 'https://api.twitch.tv/kraken/channels/'.$userId
+							CURLOPT_HTTPHEADER => [
+								'Authorization: Bearer '.$this->token,
+								'Client-ID: '.$this->config['functions_StatusTwitch']['apikay']
+							],
 						]);
-						$jdc = json_decode(curl_exec($ch));
-						$statusTwitch_name = self::$l->sprintf($value['channel_name'], '[Offline]');
-						$channel_description = self::$l->sprintf(self::$l->offline_StatusTwitch, $value['users'], $jdc->logo ?? '');
+						$games = json_decode(curl_exec($ch));
+						$strtotime = strtotime($this->streams->data[0]->started_at);
+						$data = $this->bot->przelicz_czas(time()-$strtotime);
+						$txt_time = $this->bot->wyswietl_czas($data, 1, 1, 1, 0, 0);
+						$statusTwitch_name = self::$l->sprintf($value['channel_name'], '[Online]');
+						$channel_description = self::$l->sprintf(self::$l->online_StatusTwitch, $this->users->data[0]->profile_image_url, 'https://twitch.tv/'.$this->streams->data[0]->user_name, $this->streams->data[0]->user_name, $games->data[0]->name, $this->streams->data[0]->title, $this->streams->data[0]->viewer_count, $txt_time, $this->follows->total, $this->emot);
 						$channelinfo = self::$tsAdmin->getElement('data', self::$tsAdmin->channelInfo($cid));
 						if($channelinfo['channel_description'] != $channel_description){
-							$channelEdit = self::$tsAdmin->channelEdit($cid, [ 'channel_description' => $channel_description, 'channel_name' => $statusTwitch_name ]);
+							$data['channel_description'] = $channel_description;
+							if($channelinfo['channel_name'] != $statusTwitch_name){
+								$data['channel_name'] = $statusTwitch_name;
+							}
+							$channelEdit = self::$tsAdmin->channelEdit($cid, $data);
 							if(!empty($channelEdit['errors'][0]) && $channelEdit['errors'][0] != 'ErrorID: 771 | Message: channel name is already in use'){
 								$this->bot->log(1, 'Kanał o ID:'.$cid.' nie istnieje Funkcja: statusTwitch()');
 							}
 						}
-					}else{
-						if(self::$statusTwitch_online[$userId] == 0 && $value['info'] == true){
-							self::$statusTwitch_online[$userId] = 1;
+						if(self::$statusTwitch_online[$this->users->data[0]->id] == 0 && $value['info'] == true){
+							self::$statusTwitch_online[$this->users->data[0]->id] = 1;
 							foreach($this->bot->getClientList() as $cl) {
-								self::$tsAdmin->sendMessage(1, $cl['clid'], self::$l->sprintf($value['info_text'], $value['users'], $jdc->stream->channel->url));
+								self::$tsAdmin->sendMessage(1, $cl['clid'], self::$l->sprintf($value['info_text'], $this->streams->data[0]->user_name, 'https://twitch.tv/'.$this->streams->data[0]->user_name));
 							}
 						}
-						$statusTwitch_name = self::$l->sprintf($value['channel_name'], '[Online]');
-						$channel_description = self::$l->sprintf(self::$l->online_StatusTwitch, $jdc->stream->channel->url, $value['users'], $jdc->stream->game, $jdc->stream->channel->status, $jdc->stream->viewers, $jdc->stream->channel->logo);
-						$channelEdit = self::$tsAdmin->channelEdit($cid, [ 'channel_description' => $channel_description, 'channel_name' => $statusTwitch_name ]);
-						if(!empty($channelEdit['errors'][0]) && $channelEdit['errors'][0] != 'ErrorID: 771 | Message: channel name is already in use'){
-							$this->bot->log(1, 'Kanał o ID:'.$cid.' nie istnieje Funkcja: statusTwitch()');
+					}else{
+						if(self::$statusTwitch_online[$this->users->data[0]->id] == 1){
+							self::$statusTwitch_online[$this->users->data[0]->id] = 0;
+						}
+						$statusTwitch_name = self::$l->sprintf($value['channel_name'], '[Offline]');
+						$channel_description = self::$l->sprintf(self::$l->offline_StatusTwitch, $this->users->data[0]->profile_image_url, 'https://twitch.tv/'.$this->users->data[0]->display_name, $this->users->data[0]->display_name, $this->follows->total, $this->emot);
+						$channelinfo = self::$tsAdmin->getElement('data', self::$tsAdmin->channelInfo($cid));
+						if($channelinfo['channel_description'] != $channel_description){
+							$data['channel_description'] = $channel_description;
+							if($channelinfo['channel_name'] != $statusTwitch_name){
+								$data['channel_name'] = $statusTwitch_name;
+							}
+							$channelEdit = self::$tsAdmin->channelEdit($cid, $data);
+							if(!empty($channelEdit['errors'][0]) && $channelEdit['errors'][0] != 'ErrorID: 771 | Message: channel name is already in use'){
+								$this->bot->log(1, 'Kanał o ID:'.$cid.' nie istnieje Funkcja: statusTwitch()');
+							}
 						}
 					}
 				}